chore(Structure): Rename Decorators -> Enhancers
This implements a principle we've been discussing of not using the
decorator syntax and instead favoring enhancer.
